+++ R/cache.R
@@ -240,7 +240,7 @@ renv_cache_diagnose_corrupt_metadata <- function(paths, problems, verbose) {
       renv_pretty_print(
         renv_cache_format_path(bad),
         "The following package(s) are missing 'Meta/package.rds':",
-        "These packages should be purged and re-installed.",
+        "These packages should be purged and reinstalled.",
         wrap = FALSE
       )
     }
@@ -270,7 +270,7 @@ renv_cache_diagnose_corrupt_metadata <- function(paths, problems, verbose) {
       renv_pretty_print(
         renv_cache_format_path(bad),
         "The following package(s) have corrupt 'Meta/package.rds' files:",
-        "These packages should be purged and re-installed.",
+        "These packages should be purged and reinstalled.",
         wrap = FALSE
       )
     }
@@ -302,7 +302,7 @@ renv_cache_diagnose_missing_descriptions <- function(paths, problems, verbose) {
     renv_pretty_print(
       renv_cache_format_path(bad),
       "The following packages are missing DESCRIPTION files in the cache:",
-      "These packages should be purged and re-installed.",
+      "These packages should be purged and reinstalled.",
       wrap = FALSE
     )
   }
@@ -420,7 +420,7 @@ renv_cache_diagnose_wrong_built_version <- function(paths, problems, verbose) {
     renv_pretty_print(
       renv_cache_format_path(paths[wrong]),
       "The following packages in the cache were built for a different version of R:",
-      "These packages will need to be purged and re-installed.",
+      "These packages will need to be purged and reinstalled.",
       wrap = FALSE
     )