+++ R/rebuild.R
@@ -1,7 +1,7 @@
 
 #' Rebuild the Packages in your Project Library
 #'
-#' Rebuild and re-install packages in your library. This can be useful as a
+#' Rebuild and reinstall packages in your library. This can be useful as a
 #' diagnostic tool -- for example, if you find that one or more of your
 #' packages fail to load, and you want to ensure that you are starting from a
 #' clean slate.
@@ -9,7 +9,7 @@
 #' @inherit renv-params
 #'
 #' @param packages The package(s) to be rebuilt. When `NULL`, all packages
-#'   in the library will be re-installed.
+#'   in the library will be reinstalled.
 #'
 #' @param recursive Boolean; should dependencies of packages be rebuilt
 #'   recursively? Defaults to `TRUE`.